# wiki/swagger.yaml
openapi: 3.0.0
info:
  title: 'Virta Documentation'
  contact:
    email: 'diacojocaru@icloud.com'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1'
    description: 'API Server'
paths:
...
  /api/v1/stations:
    get:
      tags:
        - stations
      summary: 'Store new station'
      description: 'Get list of stations.'
      operationId: GetStationsList
      parameters:
        -
          $ref: '#/components/parameters/PerPage'
        -
          $ref: '#/components/parameters/Page'
      responses:
        '200':
          $ref: '#/components/responses/StationResponse'
    post:
      tags:
        - stations
      summary: 'Store new station'
      description: 'Returns station data'
      operationId: StoreStation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreStationRequest'
      responses:
        '201':
          $ref: '#/components/responses/StationResponse'
        '400':
          description: 'Bad Request'
        '422':
          description: 'Unprocessable Entity.'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'The given data was invalid.' }
                  errors: { $ref: '#/components/schemas/InvalidStoreStationRequest' }
                type: object
  '/api/v1/stations/{id}':
    get:
      tags:
        - stations
      summary: 'Get station information'
      description: 'Returns station data'
      operationId: getStationById
      parameters:
        -
          name: id
          in: path
          description: 'Station id'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/StationResponse'
        '400':
          description: 'Bad Request'
        '404':
          description: 'Resource Not Found'
    put:
      tags:
        - stations
      summary: 'Update existing station'
      description: 'Returns updated station data'
      operationId: updateStation
      parameters:
        -
          name: id
          in: path
          description: 'Station id'
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStationRequest'
      responses:
        '202':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/responses/StationResponse'
        '400':
          description: 'Bad Request'
        '422':
          description: 'Unprocessable Entity.'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'The given data was invalid.' }
                  errors: { $ref: '#/components/schemas/InvalidUpdateStationRequest' }
                type: object
        '404':
          description: 'Resource Not Found'
    delete:
      tags:
        - stations
      summary: 'Delete existing station'
      description: 'Deletes a record and returns no content'
      operationId: deleteStation
      parameters:
        -
          name: id
          in: path
          description: 'Station id'
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Successful operation'
          content:
            application/json:
              schema: {  }
        '404':
          description: 'Resource Not Found'
components:
  schemas:
...
    StoreStationRequest:
      required:
        - name
        - latitude
        - longitude
        - company_id
        - address
      properties:
        name:
          type: string
          example: 'John Doe'
        latitude:
          type: number
          format: float
          example: '75.65'
        longitude:
          type: number
          format: float
          example: '120.75'
        company_id:
          type: integer
          example: '1'
        address:
          type: string
          example: '5th avenue, nex to the bank'
      type: object
    InvalidStoreStationRequest:
      properties:
        name:
          type: array
          items:
            type: string
            example:
              - 'The name field is invalid.'
        latitude:
          type: array
          items:
            type: float
            example:
              - 'The latitude field is invalid.'
        longitude:
          type: array
          items:
            type: float
            example:
              - 'The longitude field is invalid.'
        company_id:
          type: array
          items:
            type: integer
            example:
              - 'The company_id field is invalid.'
        address:
          type: array
          items:
            type: string
            example:
              - 'The address field is invalid.'
      type: object
...
    UpdateStationRequest:
      required:
        - name
        - latitude
        - longitude
        - company_id
        - address
      properties:
        name:
          type: string
          example: 'John Doe'
        latitude:
          type: number
          format: float
          example: '75.65'
        longitude:
          type: number
          format: float
          example: '120.75'
        company_id:
          type: integer
          example: '1'
        address:
          type: string
          example: '5th avenue, nex to the bank'
      type: object
    InvalidUpdateStationRequest:
      properties:
        name:
          type: array
          items:
            type: string
            example:
              - 'The name field is invalid.'
        latitude:
          type: array
          items:
            type: float
            example:
              - 'The latitude field is invalid.'
        longitude:
          type: array
          items:
            type: float
            example:
              - 'The longitude field is invalid.'
        company_id:
          type: array
          items:
            type: integer
            example:
              - 'The company_id field is invalid.'
        address:
          type: array
          items:
            type: string
            example:
              - 'The address field is invalid.'
      type: object
...
    Station:
      description: 'Transform the resource into an array.'
      required:
        - id
        - name
      properties:
        id:
          type: integer
          readOnly: true
          example: '1'
        name:
          type: string
          readOnly: false
          example: John
        latitude:
          type: number
          format: float
          readOnly: false
          example: '95.6'
        longitude:
          type: number
          format: float
          readOnly: false
          example: '97.25'
        company_id:
          type: integer
          readOnly: false
          example: '1'
        address:
          type: string
          readOnly: false
          example: 'a simple address'
      type: object
...
  responses:
...
    StationResponse:
      description: 'Station response'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Station'
...
  parameters:
...
    Page:
      name: page
      in: query
      description: 'Page number.'
      required: false
      schema:
        type: integer
    PerPage:
      name: perPage
      in: query
      description: 'number of elements per page.'
      required: false
      schema:
        type: integer
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Token based authentication.'
      bearerFormat: JWT
      scheme: bearer